+++ packages/react/src/themes/teams/components/Icon/svg/ProcessedIcons/icons-plugs.tsx
@@ -8,5 +8,5 @@ export default {
     </svg>
   ),
   styles: {},
-  exportedAs: 'connector-badge',
+  exportedAs: 'plugs',
 } as TeamsProcessedSvgIconSpec