+++ packages/react/src/themes/teams/components/Icon/svg/icons/index.ts
@@ -12,7 +12,7 @@ import ban from './ban'
 import bellMute from './bellMute'
 import bold from './bold'
 import bookmark from './bookmark'
-import bot from './bot'
+import robot from './robot'
 import broadcast from './broadcast'
 import bullets from './bullets'
 import calendar from './calendar'
@@ -40,7 +40,7 @@ import chevronRightMedium from './chevronRightMedium'
 import clipboardCopiedTo from './clipboardCopiedTo'
 import close from './close'
 import codeSnippet from './codeSnippet'
-import connectorBadge from './connectorBadge'
+import plugs from './plugs'
 import filesAftereffects from './filesAftereffects'
 import filesCode from './filesCode'
 import filesEmpty from './filesEmpty'
@@ -140,7 +140,7 @@ import star from './star'
 import stardustCircle from './stardustCircle'
 import sticker from './sticker'
 import strike from './strike'
-import tabBadge from './tabBadge'
+import tabs from './tabs'
 import table from './table'
 import tableAdd from './tableAdd'
 import tableDelete from './tableDelete'
@@ -176,7 +176,7 @@ export default {
   'bell-slash': bellMute,
   bold,
   bookmark,
-  bot,
+  robot,
   broadcast,
   bullets,
   calendar,
@@ -204,7 +204,7 @@ export default {
   'chevron-right-medium': chevronRightMedium,
   'clipboard-copied-to': clipboardCopiedTo,
   close,
-  'connector-badge': connectorBadge,
+  plugs,
   'code-snippet': codeSnippet,
   'files-aftereffects': filesAftereffects,
   'files-code': filesCode,
@@ -291,7 +291,7 @@ export default {
   star,
   sticker,
   strike,
-  'tab-badge': tabBadge,
+  tabs,
   table,
   'table-add': tableAdd,
   'table-delete': tableDelete,